requirements.txt
agate==1.9.1 ; python_version >= "3.12" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.4.0 ; python_version >= "3.12" and python_version < "4.0"
attrs==24.2.0 ; python_version >= "3.12" and python_version < "4.0"
babel==2.16.0 ; python_version >= "3.12" and python_version < "4.0"
cachetools==5.5.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.8.30 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0"
daff==1.3.46 ; python_version >= "3.12" and python_version < "4.0"
db-dtypes==1.3.0 ; python_version >= "3.12" and python_version < "4.0"
dbt-adapters==1.4.1 ; python_version >= "3.12" and python_version < "4.0"
dbt-bigquery==1.8.2 ; python_version >= "3.12" and python_version < "4.0"
dbt-common==1.7.0 ; python_version >= "3.12" and python_version < "4.0"
dbt-core==1.8.6 ; python_version >= "3.12" and python_version < "4.0"
dbt-extractor==0.5.1 ; python_version >= "3.12" and python_version < "4.0"
dbt-semantic-interfaces==0.5.1 ; python_version >= "3.12" and python_version < "4.0"
deepdiff==7.0.1 ; python_version >= "3.12" and python_version < "4.0"
fastapi==0.112.2 ; python_version >= "3.12" and python_version < "4.0"
google-api-core==2.19.2 ; python_version >= "3.12" and python_version < "4.0"
google-api-core[grpc]==2.19.2 ; python_version >= "3.12" and python_version < "4.0"
google-auth==2.34.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-bigquery[pandas]==3.25.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-core==2.4.1 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-dataproc==5.10.2 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-storage==2.18.2 ; python_version >= "3.12" and python_version < "4.0"
google-crc32c==1.5.0 ; python_version >= "3.12" and python_version < "4.0"
google-resumable-media==2.7.2 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos==1.65.0 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos[grpc]==1.65.0 ; python_version >= "3.12" and python_version < "4.0"
grpc-google-iam-v1==0.13.1 ; python_version >= "3.12" and python_version < "4.0"
grpcio-status==1.62.3 ; python_version >= "3.12" and python_version < "4.0"
grpcio==1.66.1 ; python_version >= "3.12" and python_version < "4.0"
gunicorn==23.0.0 ; python_version >= "3.12" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
httptools==0.6.1 ; python_version >= "3.12" and python_version < "4.0"
idna==3.8 ; python_version >= "3.12" and python_version < "4.0"
importlib-metadata==6.11.0 ; python_version >= "3.12" and python_version < "4.0"
isodate==0.6.1 ; python_version >= "3.12" and python_version < "4.0"
jinja2==3.1.4 ; python_version >= "3.12" and python_version < "4.0"
jsonschema-specifications==2023.12.1 ; python_version >= "3.12" and python_version < "4.0"
jsonschema==4.23.0 ; python_version >= "3.12" and python_version < "4.0"
leather==0.4.0 ; python_version >= "3.12" and python_version < "4.0"
logbook==1.5.3 ; python_version >= "3.12" and python_version < "4.0"
loguru==0.7.2 ; python_version >= "3.12" and python_version < "4.0"
markupsafe==2.1.5 ; python_version >= "3.12" and python_version < "4.0"
mashumaro[msgpack]==3.13.1 ; python_version >= "3.12" and python_version < "4.0"
minimal-snowplow-tracker==0.0.2 ; python_version >= "3.12" and python_version < "4.0"
more-itertools==10.4.0 ; python_version >= "3.12" and python_version < "4.0"
msgpack==1.0.8 ; python_version >= "3.12" and python_version < "4.0"
networkx==3.3 ; python_version >= "3.12" and python_version < "4.0"
numpy==2.1.0 ; python_version >= "3.12" and python_version < "4.0"
ordered-set==4.1.0 ; python_version >= "3.12" and python_version < "4.0"
packaging==24.1 ; python_version >= "3.12" and python_version < "4.0"
pandas==2.2.2 ; python_version >= "3.12" and python_version < "4.0"
parsedatetime==2.6 ; python_version >= "3.12" and python_version < "4.0"
pathspec==0.12.1 ; python_version >= "3.12" and python_version < "4.0"
proto-plus==1.24.0 ; python_version >= "3.12" and python_version < "4.0"
protobuf==4.25.4 ; python_version >= "3.12" and python_version < "4.0"
pyarrow==17.0.0 ; python_version >= "3.12" and python_version < "4.0"
pyasn1-modules==0.4.0 ; python_version >= "3.12" and python_version < "4.0"
pyasn1==0.6.0 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.20.1 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.8.2 ; python_version >= "3.12" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
python-slugify==8.0.4 ; python_version >= "3.12" and python_version < "4.0"
pytimeparse==1.1.8 ; python_version >= "3.12" and python_version < "4.0"
pytz==2024.1 ; python_version >= "3.12" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.12" and python_version < "4.0"
referencing==0.35.1 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
rpds-py==0.20.0 ; python_version >= "3.12" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.12" and python_version < "4"
six==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
sqlparse==0.5.1 ; python_version >= "3.12" and python_version < "4.0"
starlette==0.38.4 ; python_version >= "3.12" and python_version < "4.0"
text-unidecode==1.3 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.2.2 ; python_version >= "3.12" and python_version < "4.0"
uvicorn[standard]==0.30.6 ; python_version >= "3.12" and python_version < "4.0"
uvloop==0.20.0 ; (sys_platform != "win32" and sys_platform != "cygwin") and platform_python_implementation != "PyPy" and python_version >= "3.12" and python_version < "4.0"
watchfiles==0.24.0 ; python_version >= "3.12" and python_version < "4.0"
websockets==13.0.1 ; python_version >= "3.12" and python_version < "4.0"
win32-setctime==1.1.0 ; python_version >= "3.12" and python_version < "4.0" and sys_platform == "win32"
zipp==3.20.1 ; python_version >= "3.12" and python_version < "4.0"
